refactor in smaller components
